requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
deepface
opencv-python-headless
mediapipe
streamlit
sentence-transformers
faiss-cpu
google-generativeai
pydantic
python-multipart
tf-keras
pydantic-settings
Pillow
websockets
edge-tts
twilio
transformers
torch
torchaudio
openai
fpdf2
